requirements.txt
# Core dependencies
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Data handling
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Date handling
python-dateutil==2.8.2

# Reporting
reportlab==4.0.7
jinja2==3.1.2

# CLI interface
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
